feat(imageqa): visual qa and doc qa fixes
